Added $saveLocation Parameter
By adding this parameter fixes a problem where it was saving the file
in the folder where the module was saved.  Now you an specify a
location to save the module or if you leave it blank it will save the
file into the working directory of where the function is being called
from.