Added Progress Bar to New-StaffCalendar Function
Added progress bar when the New-StaffCalendar function is running so
that there is a visual that the script is running and working
correctly and not hung because the script takes about a minute to
create the excel file.  This also allows me to have the temp excel
file that had all the changes being done to it to be hidden.